feat(reference): fold Pen Compat + Driver Compat into Reference sections
Move the two standalone Data sub-nav routes (/pen-compat,
/wacom-driver-compat) into the Reference page as sections under a new
"Compatibility" category, so all reference-style data lives in one place.

- reference/+page.ts merges the three loaders: builds enriched pen-compat
  rows (TabletFullName/PenFullName via canonical formatters, brand-id
  fallback for orphans) plus the WACOM model/sensor-id maps for the driver
  join.
- reference/+page.svelte adds two sectionDefs and content branches: an
  embedded EntityExplorer for pen compatibility (hash vs query-param URL
  state don't collide) and a new DriverCompatSection component for the
  Wacom update.xml → tablet join.
- DriverCompatSection.svelte: extracted driver-compat table with the
  sensorid-then-dashless-model match logic and EntityLink-ed matched rows.
- Drop both tabs from dataSubNavTabs(); trim Nav Data altActive; delete the
  two routes; data-dictionary uses dataSubNavTabs() (was a stale hardcoded
  list still naming the removed routes).
- smoke.spec.ts: drop the deleted /pen-compat route entry (covered by
  /reference).

check 0/0, lint 0 errors, 558 unit, 27 e2e green.
