+++ src/lib/components/Nav.svelte
@@ -35,14 +35,7 @@
 		{
 			href: '/reference',
 			label: 'Data',
-			altActive: [
-				'/brands',
-				'/data-dictionary',
-				'/api-explorer',
-				'/data-quality',
-				'/pen-compat',
-				'/wacom-driver-compat',
-			],
+			altActive: ['/brands', '/data-dictionary', '/api-explorer', '/data-quality'],
 		},
 		{ href: '/about', label: 'About' },
 	];
