+++ src/lib/nav/subnav-tabs.ts
@@ -25,8 +25,6 @@ export function dataSubNavTabs(): SubNavTab[] {
 		{ href: '/data-dictionary', label: 'Data Dictionary' },
 		{ href: '/api-explorer', label: 'API Explorer' },
 		{ href: '/data-quality', label: 'Data Quality' },
-		{ href: '/pen-compat', label: 'Pen Compat' },
-		{ href: '/wacom-driver-compat', label: 'Driver Compat' },
 	];
 }
